Setting up test cases for Rust Linked List
Learning some rust syntax, it is pretty cool. Also re-doing some of how I want to do this project, as having the data structures in multiple languages is going to be crazy redundant (at least for simple ones). In the future, I might do practice problems or something or switch up the languages as I go.